- student list screen design and navigation - test repository

+++ app/src/main/java/com/example/happystudent/core/database/model/StudentEntity.kt
@@ -9,7 +9,7 @@ data class StudentEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     val name: String,
-    val leaving_probability: Float,
+    val leaving_probability: Double,
     val group: String,
     val update_date: String,
     val imageUrl: String
