+++ app/src/main/java/com/example/happystudent/core/model/Student.kt
@@ -3,7 +3,7 @@ package com.example.happystudent.core.model
 data class Student(
     val id: Int = UNDEFINED_ID,
     val name: String,
-    val leaving_probability: Float,
+    val leaving_probability: Double,
     val group: String,
     val update_date: String,
     val imageUrl: String
